+++ cloudauth-intl-20220809/alibabacloud_cloudauth_intl20220809/models/_face_compare_v2advance_request.py
@@ -10,6 +10,7 @@ class FaceCompareV2AdvanceRequest(DaraModel):
     def __init__(
         self,
         face_picture_quality_check: str = None,
+        face_quality_check: str = None,
         merchant_biz_id: str = None,
         source_face_picture: str = None,
         source_face_picture_file_object: BinaryIO = None,
@@ -19,6 +20,7 @@ class FaceCompareV2AdvanceRequest(DaraModel):
         target_face_picture_url: str = None,
     ):
         self.face_picture_quality_check = face_picture_quality_check
+        self.face_quality_check = face_quality_check
         self.merchant_biz_id = merchant_biz_id
         self.source_face_picture = source_face_picture
         self.source_face_picture_file_object = source_face_picture_file_object
@@ -37,6 +39,9 @@ class FaceCompareV2AdvanceRequest(DaraModel):
             result = _map
         if self.face_picture_quality_check is not None:
             result['FacePictureQualityCheck'] = self.face_picture_quality_check
+
+        if self.face_quality_check is not None:
+            result['FaceQualityCheck'] = self.face_quality_check
 
         if self.merchant_biz_id is not None:
             result['MerchantBizId'] = self.merchant_biz_id
@@ -65,6 +70,9 @@ class FaceCompareV2AdvanceRequest(DaraModel):
         m = m or dict()
         if m.get('FacePictureQualityCheck') is not None:
             self.face_picture_quality_check = m.get('FacePictureQualityCheck')
+
+        if m.get('FaceQualityCheck') is not None:
+            self.face_quality_check = m.get('FaceQualityCheck')
 
         if m.get('MerchantBizId') is not None:
             self.merchant_biz_id = m.get('MerchantBizId')