+++ cloudauth-intl-20220809/alibabacloud_cloudauth_intl20220809/models/_face_compare_v2response_body.py
@@ -62,22 +62,28 @@ class FaceCompareV2ResponseBody(DaraModel):
 class FaceCompareV2ResponseBodyResult(DaraModel):
     def __init__(
         self,
+        ext_face_info: main_models.FaceCompareV2ResponseBodyResultExtFaceInfo = None,
         face_comparison_score: float = None,
         passed: str = None,
         transaction_id: str = None,
     ):
+        self.ext_face_info = ext_face_info
         self.face_comparison_score = face_comparison_score
         self.passed = passed
         self.transaction_id = transaction_id
 
     def validate(self):
-        pass
+        if self.ext_face_info:
+            self.ext_face_info.validate()
 
     def to_map(self):
         result = dict()
         _map = super().to_map()
         if _map is not None:
             result = _map
+        if self.ext_face_info is not None:
+            result['ExtFaceInfo'] = self.ext_face_info.to_map()
+
         if self.face_comparison_score is not None:
             result['FaceComparisonScore'] = self.face_comparison_score
 
@@ -91,6 +97,10 @@ class FaceCompareV2ResponseBodyResult(DaraModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('ExtFaceInfo') is not None:
+            temp_model = main_models.FaceCompareV2ResponseBodyResultExtFaceInfo()
+            self.ext_face_info = temp_model.from_map(m.get('ExtFaceInfo'))
+
         if m.get('FaceComparisonScore') is not None:
             self.face_comparison_score = m.get('FaceComparisonScore')
 
@@ -99,6 +109,65 @@ class FaceCompareV2ResponseBodyResult(DaraModel):
 
         if m.get('TransactionId') is not None:
             self.transaction_id = m.get('TransactionId')
+
+        return self
+
+class FaceCompareV2ResponseBodyResultExtFaceInfo(DaraModel):
+    def __init__(
+        self,
+        face_quality_score: float = None,
+        illumination_score: float = None,
+        ka_occlusion_score: float = None,
+        occlusion_score: float = None,
+        sharpness_score: float = None,
+    ):
+        self.face_quality_score = face_quality_score
+        self.illumination_score = illumination_score
+        self.ka_occlusion_score = ka_occlusion_score
+        self.occlusion_score = occlusion_score
+        self.sharpness_score = sharpness_score
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.face_quality_score is not None:
+            result['FaceQualityScore'] = self.face_quality_score
+
+        if self.illumination_score is not None:
+            result['IlluminationScore'] = self.illumination_score
+
+        if self.ka_occlusion_score is not None:
+            result['KaOcclusionScore'] = self.ka_occlusion_score
+
+        if self.occlusion_score is not None:
+            result['OcclusionScore'] = self.occlusion_score
+
+        if self.sharpness_score is not None:
+            result['SharpnessScore'] = self.sharpness_score
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('FaceQualityScore') is not None:
+            self.face_quality_score = m.get('FaceQualityScore')
+
+        if m.get('IlluminationScore') is not None:
+            self.illumination_score = m.get('IlluminationScore')
+
+        if m.get('KaOcclusionScore') is not None:
+            self.ka_occlusion_score = m.get('KaOcclusionScore')
+
+        if m.get('OcclusionScore') is not None:
+            self.occlusion_score = m.get('OcclusionScore')
+
+        if m.get('SharpnessScore') is not None:
+            self.sharpness_score = m.get('SharpnessScore')
 
         return self